Move eeglab test to eeglab
